push

+++ app/Http/Controllers/Theme/HomeController.php
@@ -36,6 +36,7 @@ use Illuminate\Support\Str;
 use \Cart as Cart;
 use Session;
 use App\Services\FBPixelService;
+use App\Model\WaitingList;
 
 class HomeController extends Controller
 {
@@ -402,6 +403,109 @@ class HomeController extends Controller
             }
         }
         
+        Cart::instance('default')->destroy();
+        Session::forget('pay_seats_data');
+        Session::forget('transaction_id');
+        Session::forget('cardtype');
+        Session::forget('installments');
+        //Session::forget('pay_invoice_data');
+        Session::forget('pay_bill_data');
+        Session::forget('deree_user_data');
+        Session::forget('user_id');
+        Session::forget('coupon_code');
+        Session::forget('coupon_price');
+        Session::forget('priceOf');
+
+        $data['info']['success'] = true;
+        $data['info']['title'] = __('thank_you_page.title');
+        $data['info']['message'] = __('thank_you_page.message');
+        $data['info']['statusClass'] = 'success';
+
+        $data['event']['title'] = $content->title;
+        $data['event']['slug'] = $content->slugable->slug;
+        $data['event']['facebook'] = url('/') . '/' .$content->slugable->slug .'?utm_source=Facebook&utm_medium=Post_Student&utm_campaign=KNOWCRUNCH_BRANDING&quote='.urlencode("Proudly participating in ". $content->title . " by Knowcrunch.");
+        $data['event']['twitter'] = urlencode("Proudly participating in ". $content->title . " by Knowcrunch. 💙");
+        $data['event']['linkedin'] = urlencode(url('/') . '/' .$content->slugable->slug .'?utm_source=LinkedIn&utm_medium=Post_Student&utm_campaign=KNOWCRUNCH_BRANDING&title='."Proudly participating in ". $content->title . " by Knowcrunch. 💙");
+        
+        Session::put('thankyouData',$data);
+        return redirect('/thankyou');
+        //return view('theme.cart.new_cart.thank_you_free',$data);
+
+    }
+
+    public function enrollToWaitingList(Event $content){
+        
+        $published = $content->published;
+        $status = $content->status;
+        if($published == 0 || $status != 5){
+            return false;
+        }
+        
+
+        
+        $data['user']['createAccount'] = false;
+
+        if( !($user = Auth::user()) && !($user = User::where('email',request()->email[0])->first()) ){
+
+            $data['user']['createAccount'] = true;
+
+            $input = [];
+            $formData = request()->all();
+            unset($formData['_token']);
+            unset($formData['terms_condition']);
+            unset($formData['update']);
+            unset($formData['type']);
+            
+            foreach($formData as $key => $value){
+                $input[$key] = $value[0];
+            }
+
+            $input['password'] = Hash::make(date('Y-m-dTH:i:s'));
+            
+            $user = User::create($input);
+
+            $code = Activation::create([
+                'user_id' => $user->id,
+                'code' => Str::random(40),
+                'completed' => true,
+            ]);
+            $user->role()->attach(7);
+
+
+
+            $cookieValue = base64_encode($user->id . date("H:i"));
+            setcookie('auth-'.$user->id, $cookieValue, time() + (1 * 365 * 86400), "/"); // 86400 = 1 day
+        
+            $coockie = new CookiesSMS;
+            $coockie->coockie_name = 'auth-'.$user->id;
+            $coockie->coockie_value = $cookieValue;
+            $coockie->user_id = $user->id;
+            $coockie->sms_code = -1;
+            $coockie->sms_verification = true;
+
+            $coockie->save();
+        }
+
+        if(WaitingList::where('user_id',$user->id)->where('event_id',$content->id)->first()){
+            return false;
+        }
+
+        $waiting = new WaitingList;
+        $waiting->user_id = $user->id;
+        $waiting->event_id = $content->id;
+        $waiting->save();
+
+        $data['user']['first'] = $user->firstname;
+        $data['user']['name'] = $user->firstname . ' ' . $user->lastname;
+        $data['user']['email'] = $user->email;
+        $data['extrainfo'] = ['','',$content->title];
+        $data['duration'] =  $content->summary1->where('section','date')->first() ? $content->summary1->where('section','date')->first()->title : '';
+
+        $data['eventSlug'] =  url('/') . '/' . $content->getSlug();
+        //$user->notify(new WelcomeEmail($user,$data));
+
+        
+       
         Cart::instance('default')->destroy();
         Session::forget('pay_seats_data');
         Session::forget('transaction_id');
